+++ mvs_eland_tool/mvs_eland_tool.py
@@ -28,7 +28,6 @@ child-sub:  Sub-child function, feeds only back to child functions
 """
 
 import logging
-import os
 
 # Loading all child functions
 
@@ -60,6 +59,10 @@ def main(**kwargs):
         + "\n Coded at: Reiner Lemoine Institute (Berlin) "
         + "\n Contributors: Martha M. Hoffmann \n \n "
     )
+    # Parse the arguments from the command line
+    parser = initializing.create_parser()
+    args = vars(parser.parse_args())
+    kwargs.update(**args)
 
     logging.debug("Accessing script: A0_initialization")
     user_input = initializing.welcome(welcome_text, **kwargs)
@@ -69,12 +72,13 @@ def main(**kwargs):
     #    # todo: is user input completely used?
     #    dict_values = data_input.load_json(user_input["path_input_file"])
 
-    logging.debug("Accessing script: A1_csv_to_json")
-    path_to_json_from_csv = load_data_from_csv.create_input_json()
-    user_input.update({"path_input_file": path_to_json_from_csv})
+    if not kwargs["path_input_file"].endswith("json"):
+        logging.debug("Accessing script: A1_csv_to_json")
+        path_to_json_from_csv = load_data_from_csv.create_input_json()
+        user_input.update({"path_input_file": path_to_json_from_csv})
 
     logging.debug("Accessing script: B0_data_input_json")
-    dict_values = data_input.load_json(user_input)
+    dict_values = data_input.load_json(user_input["path_input_file"])
 
     print("")
     logging.debug("Accessing script: C0_data_processing")